fix weight and cost (were ints become doubles)

+++ src/main/java/com/example/task/dto/Animal.java
@@ -6,8 +6,8 @@ public class Animal{
     private String name;
     private String type;
     private String sex;
-    private Integer weight;
-    private Integer cost;
+    private Double weight;
+    private Double cost;
     public Integer determineCategory(){
         if (cost <= 20) {
             return Category.FIRST.getValue();
